internal-research-agent/requirements.txt
langchain>=0.1.1
langchain-community>=0.0.20
langchain-google-genai>=0.0.6
langchain-chroma>=0.2.4
langchain-tavily>=0.1.0
langchain-openai>=0.0.5
langchain-mcp-adapters>=0.1.0
langgraph>=0.1.0
chromadb>=0.4.22
tavily-python>=0.3.1
python-dotenv>=1.0.0
click>=8.1.7
rich>=13.7.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
unstructured>=0.11.8
python-docx>=1.1.0
PyPDF2>=3.0.1
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
opentelemetry-proto==1.34.1
mcp>=1.0.0